added masked convolutional class
